🚚 (types): Rename `httpErrors` dir to `httpResponses`
