+++ packages/types/src/httpResponses/ErrorResponse.ts
